# web/openapi.yaml
openapi: 3.1.0
info:
  title: SkillHub API
  description: |
    SkillHub registry HTTP API — discovery, publishing, authentication, namespaces,
    stars/comments/ratings, notifications, and admin operations.

    All JSON endpoints live under `/api/v1`. Bearer tokens (`Authorization: Bearer <token>`)
    are obtained via the device flow (`/api/v1/auth/device/*`) or via a logged-in web
    session cookie (`session_token`).

    See `/skills.md` for a runnable curl-based walkthrough designed for AI agents.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: /
    description: Same-origin
...
paths:
...
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: List skills (paginated)
      parameters:
        - { name: cursor,   in: query, schema: { type: string },  description: "Opaque cursor from previous nextCursor" }
        - { name: sort,     in: query, schema: { type: string, enum: [created, downloads, stars, updated], default: created } }
        - { name: limit,    in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Paginated skill list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: "#/components/schemas/Skill" } }
                  nextCursor: { type: string, nullable: true }
    post:
      tags: [Skills]
      summary: Publish a new skill version (multipart)
      description: |
        At least one file is required and `SKILL.md` must be present in the upload.
        With a team token, the `namespace` field MUST equal the token's bound namespace.
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [slug, version, files]
              properties:
                slug:         { type: string, description: "Unique skill slug" }
                version:      { type: string, description: "Semver, e.g. 0.1.0" }
                displayName:  { type: string }
                summary:      { type: string }
                category:     { type: string }
                kind:         { type: string }
                visibility:   { type: string, enum: [public, private] }
                namespace:    { type: string, description: "Namespace slug (required when using a team token)" }
                changelog:    { type: string }
                tags:         { type: string, description: "Comma- or semicolon-separated" }
                dependencies: { type: string, description: "JSON array of {slug, versionRange}" }
                signature:    { type: string, description: "Cosign bundle, max 64 KB" }
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Skill + new version
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill:   { $ref: "#/components/schemas/Skill" }
                  version: { $ref: "#/components/schemas/SkillVersion" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/skills/{slug}:
    parameters:
      - { name: slug, in: path, required: true, schema: { type: string } }
    get:
      tags: [Skills]
      summary: Get a skill by slug
      description: |
        Returns 409 with a `candidates` list when the bare slug exists in
        multiple namespaces. Use `/api/v1/skills/@{namespace}/{slug}` to
        disambiguate.
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Skill" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/AmbiguousSlug" }
    delete:
      tags: [Skills]
      summary: Soft-delete a skill (owner)
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/MessageOK" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/AmbiguousSlug" }
...
  /api/v1/admin/skills:
    get:
      tags: [Admin]
      summary: List all skills (any visibility)
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - { name: visibility, in: query, schema: { type: string, enum: [public, private, pending] } }
        - { name: cursor,     in: query, schema: { type: string } }
        - { name: limit,      in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Skills
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: "#/components/schemas/Skill" } }
                  nextCursor: { type: string, nullable: true }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Personal or team token from the device flow / `/api/v1/tokens`. Format is
        opaque; treat it as a secret string. Pass as `Authorization: Bearer <token>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Set by `POST /login` and the OAuth callback. Browser-only.
  responses:
    BadRequest:
      description: Validation error.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid token.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authenticated but not allowed.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    AmbiguousSlug:
      description: |
        The bare slug matches skills in multiple namespaces. Re-issue the
        request with a namespace-qualified path: `/api/v1/skills/@<namespace>/<slug>`.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:      { type: string, example: "ambiguous slug 'my-skill': exists in multiple namespaces" }
              candidates:
                type: array
                items: { $ref: "#/components/schemas/AmbiguousCandidate" }
    MessageOK:
      description: Success.
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
  schemas:
    AmbiguousCandidate:
      type: object
      properties:
        namespace:   { type: string, description: "Namespace slug" }
        slug:        { type: string }
        ownerHandle: { type: string }
        skillId:     { type: string, format: uuid }
    Error:
      type: object
      properties:
        error: { type: string }
...
    Skill:
      type: object
      properties:
        id:            { type: string, format: uuid }
        slug:          { type: string }
        displayName:   { type: string }
        summary:       { type: string }
        description:   { type: string }
        category:      { type: string, nullable: true }
        kind:          { type: string, nullable: true }
        tags:          { type: array, items: { type: string } }
        visibility:    { type: string, enum: [public, private] }
        latestVersion: { type: string }
        downloads:     { type: integer }
        stars:         { type: integer }
        ownerId:       { type: string, format: uuid }
        namespaceId:   { type: string, format: uuid, nullable: true }
        namespaceSlug: { type: string, description: "Slug of the namespace this skill belongs to" }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
    SkillVersion:
      type: object
      properties:
        id:              { type: string, format: uuid }
        skillId:         { type: string, format: uuid }
        version:         { type: string, description: "Semver" }
        changelog:       { type: string }
        fingerprint:     { type: string, description: "SHA-256 hex" }
        signatureBundle: { type: string, nullable: true }
        yankedAt:        { type: string, format: date-time, nullable: true }
        yankReason:      { type: string, nullable: true }
        deprecatedAt:    { type: string, format: date-time, nullable: true }
        deprecateMessage: { type: string, nullable: true }
        dependencies:
          type: array
          items:
            type: object
            properties:
              slug:         { type: string }
              versionRange: { type: string }
        publishedAt:     { type: string, format: date-time }
...